Refactor product management by introducing a custom hook for fetching and deleting products
